pyproject.toml
[project]
name = "playwright-Api-codegen"
version = "0.1.0"
description = "Rest Api Swagger to Playwright REST API Code generation."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "PyYAML>=6.0",
    "openapi-schema-pydantic",
    "python-dotenv",
    "prance",
    "openapi-spec-validator",
    "jsonref",
    "langchain>=0.2.0",
    "langchain-google-genai>=0.1.0",
    "langchain-community",
    "openai",
    "setuptools>=61.0",
    "wheel",
    "build",
    "DeepDiff"
]

[build-system]
requires = ["setuptools>=61.0", "wheel", "build"]
build-backend = "setuptools.build_meta"

[project.scripts]
playwright-restapi-swagger = "RestPlaywright.main:main"
